tests: move test_openFile() test to apps/ui-tests-app
Small changes to openFile() method
