+++ apps/app/App_Resources/Android/AndroidManifest.xml
@@ -27,6 +27,17 @@
 		android:theme="@style/AppTheme" 
 		android:usesCleartextTraffic="true" >
 		<meta-data android:name="debugLayouts" android:value="true" />
+
+		<provider
+			android:name="android.support.v4.content.FileProvider"
+			android:authorities="org.nativescript.apps.provider"
+			android:exported="false"
+			android:grantUriPermissions="true">
+		<meta-data
+			android:name="android.support.FILE_PROVIDER_PATHS"
+			android:resource="@xml/provider_paths"/>
+		</provider>
+
 		<activity
 			android:name="com.tns.NativeScriptActivity"
 			android:label="@string/title_activity_kimera"
