+++ apps/app/App_Resources/Android/xml/provider_paths.xml
