+++ apps/app/ui-tests-app/intent/open-file.xml
@@ -1,0 +1,5 @@
+<Page navigatingTo="navigatingTo">
+  <GridLayout id="root">
+    <Button text="Open File" tap="openFile" height="40"></Button>
+  </GridLayout>
+</Page>
